requirements.txt
flask==3.0.0
numpy==1.24.3
faiss-cpu==1.7.4
transformers==4.35.2
requests==2.31.0
python-dotenv==1.0.0
datasets==2.14.6
huggingface-hub==0.19.4
tokenizers==0.15.0
tqdm==4.66.1
gunicorn>=21.2.0
Jinja2==3.1.2
scikit-learn==1.3.2
scipy==1.11.4
pandas==2.1.3
